+++ src/Attribute/AsOption.php
@@ -8,15 +8,16 @@ use Symfony\Component\Console\Completion\CompletionInput;
 class AsOption extends AsCommandArgument
 {
     /**
-     * @param string|array<string>|null                      $shortcut
-     * @param array<string>                                  $suggestedValues
-     * @param mixed|callable(CompletionInput): array<string> $autocomplete
+     * @param string|array<string>|null                                    $shortcut
+     * @param array<string>                                                $suggestedValues
+     * @param mixed|array<string>|callable(CompletionInput): array<string> $autocomplete
      */
     public function __construct(
         ?string $name = null,
         public readonly string|array|null $shortcut = null,
         public readonly ?int $mode = null,
         public readonly string $description = '',
+        /** @deprecated since Castor 0.18, use "autocomplete" property instead */
         public readonly array $suggestedValues = [],
         public readonly mixed $autocomplete = null,
     ) {