+++ src/Console/Command/TaskCommand.php
@@ -242,12 +242,21 @@ class TaskCommand extends Command implements SignalableCommandInterface
      */
     private function getSuggestedValues(AsArgument|AsOption $attribute): array|\Closure
     {
+        if ($attribute->suggestedValues && !\in_array('_complete', $_SERVER['argv'], true)) {
+            // Only trigger deprecation when not in completion mode
+            trigger_deprecation('jolicode/castor', '0.18', \sprintf('The "suggestedValues" property of attribute "%s" is deprecated, use "autocomplete" property instead.', self::class));
+        }
+
         if ($attribute->suggestedValues && null !== $attribute->autocomplete) {
             throw new FunctionConfigurationException(\sprintf('You cannot define both "suggestedValues" and "autocomplete" option on parameter "%s".', $attribute->name), $this->taskDescriptor->function);
         }
 
         if (null === $attribute->autocomplete) {
             return $attribute->suggestedValues;
+        }
+
+        if (\is_array($attribute->autocomplete)) {
+            return $attribute->autocomplete;
         }
 
         if (!\is_callable($attribute->autocomplete)) {